feat: :boom: Completely recode the entire project
The entire magnesium project has been recoded from the ground up, with new features and dependencys. This is a major, breaking change and old installations will no longer work, this is to prepare us for the future of magnesium!

BREAKING CHANGES: All previous instances will break because of this recode.